+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for ASP.NET Core-compatible key format
 - Removed the `cleanup` command
+- Fail if environment in JSON secret has invalid keys
 
 ## 0.1.0 (2021-01-03)